requirements.txt
selenium==4.15.0
Appium-Python-Client==3.10.1
pytest==7.4.0
pytest-xdist==3.3.0
requests==2.31.0
pandas==2.2.3
allure-pytest==2.13.8
